+++ site/changelog.md
@@ -1,6 +1,30 @@
 ## 更新日志
 
 
+### 2.0.2-beta.12
+
+* **[feat]**:
+    - feature(input): Input组件透传tooltips配置项
+
+
+### 2.0.2-beta.11
+
+* **[feat]**:
+    - bugfix(tree): tree树组件缩进问题&异步加载子节点触发事件配置支持
+
+
+### 2.0.2-beta.10
+
+* **[feat]**:
+    - bugfix(slider): 兼容slider oldValue 为空的情况
+
+
+### 2.0.2-beta.9
+
+* **[feat]**:
+    - bugfix(tree): tree树组件缩进问题
+
+
 ### 2.0.2-beta.7
 
 * **[feat]**: